Move otel plugin
