+++ litestar/contrib/opentelemetry/__init__.py
@@ -1,4 +1,9 @@
 from .config import OpenTelemetryConfig
 from .middleware import OpenTelemetryInstrumentationMiddleware
+from .plugin import OpenTelemetryPlugin
 
-__all__ = ("OpenTelemetryConfig", "OpenTelemetryInstrumentationMiddleware")
+__all__ = (
+    "OpenTelemetryConfig",
+    "OpenTelemetryInstrumentationMiddleware",
+    "OpenTelemetryPlugin",
+)
